requirements.txt
# ClearPath Chatbot Backend - Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.1.0
PyMuPDF==1.25.4
rank-bm25==0.2.2
numpy>=1.24.0
groq==0.25.0
pydantic>=2.0.0
